consumer/requirements.txt
kafka-python
grpcio
psycopg2-binary